requirements.txt
instabot==0.28.0
python-dotenv==0.10.1

